libzdb: increase fileid_t from 16 bits to 32 bits
Warning: major breaking change.

Move from 16 bits fileid to 32 bits fileid.

The 16 bits was limiting amount of files to 65536.
In early version, datafile size were expected to be like minimum
128 or 256 MB, which was fine to go up to couple of TB per namespace.

In new version, there are new usecase where smaller datafile are
required and even dynamic datasize (rotation before maximum size
reached) which would lead to lot of smaller files. In this case, 65k
files will be really low and would set a hard limit to namespace size.

Eg: 8 MB datafile would lead to maximum 524 GB maximum namespace size.

This type change break internal file specification and thus are not
backward compatible. But this change move the limitation to nearly no
limitation at all.

New change allows 8 MB datasize to maximum ~32 PB

+++ libzdb/libzdb.h
@@ -118,7 +118,7 @@
     // fileid_t represent internal size of fileid type, which was
     // historically hardcoded to uint16_t everywhere, but make code less
     // easy to update if this size need to change
-    typedef uint16_t fileid_t;
+    typedef uint32_t fileid_t;
 
     void zdb_tools_fulldump(void *_data, size_t len);
     void zdb_tools_hexdump(void *input, size_t length);
